+++ .gitignore
@@ -8,3 +8,6 @@ node_modules
 .idea/
 *.log
 cypress/downloads
+
+# mac temp file
+._*
