+++ package.json
@@ -20,23 +20,23 @@
   },
   "dependencies": {
     "@element-plus/icons-vue": "^2.3.1",
-    "@unhead/vue": "^1.8.9",
-    "@unocss/reset": "^0.58.1",
-    "@vueuse/core": "^10.7.0",
+    "@unhead/vue": "^1.9.4",
+    "@unocss/reset": "^0.58.9",
+    "@vueuse/core": "^10.9.0",
     "@vueuse/head": "^2.0.0",
-    "element-plus": "^2.5.6",
+    "element-plus": "^2.6.3",
     "nprogress": "^0.2.0",
     "pinia": "^2.1.7",
-    "sass": "^1.71.1",
-    "swiper": "^11.0.7",
-    "vue": "^3.3.13",
-    "vue-demi": "^0.14.6",
-    "vue-i18n": "^9.8.0",
-    "vue-router": "^4.2.5"
+    "sass": "^1.74.1",
+    "swiper": "^11.1.0",
+    "vue": "^3.4.21",
+    "vue-demi": "^0.14.7",
+    "vue-i18n": "^9.11.0",
+    "vue-router": "^4.3.0"
   },
   "devDependencies": {
-    "@antfu/eslint-config": "^2.6.1",
-    "@iconify-json/carbon": "^1.1.27",
+    "@antfu/eslint-config": "^2.12.1",
+    "@iconify-json/carbon": "^1.1.31",
     "@intlify/unplugin-vue-i18n": "^2.0.0",
     "@storybook/addon-essentials": "^7.6.17",
     "@storybook/addon-interactions": "^7.6.17",
@@ -48,48 +48,48 @@
     "@storybook/vue3-vite": "^7.6.17",
     "@types/markdown-it-link-attributes": "^3.0.4",
     "@types/nprogress": "^0.2.3",
-    "@unocss/eslint-config": "^0.58.1",
-    "@unocss/vite": "^0.58.5",
-    "@vitejs/plugin-vue": "^5.0.0",
-    "@vue-macros/volar": "^0.18.8",
-    "@vue/test-utils": "^2.4.3",
+    "@unocss/eslint-config": "^0.58.9",
+    "@unocss/vite": "^0.58.9",
+    "@vitejs/plugin-vue": "^5.0.4",
+    "@vue-macros/volar": "^0.18.17",
+    "@vue/test-utils": "^2.4.5",
     "critters": "^0.0.20",
     "cross-env": "^7.0.3",
-    "cypress": "^13.6.2",
+    "cypress": "^13.7.2",
     "cypress-vite": "^1.5.0",
-    "eslint": "^8.56.0",
+    "eslint": "^8.57.0",
     "eslint-plugin-cypress": "^2.15.1",
     "eslint-plugin-format": "^0.1.0",
     "eslint-plugin-storybook": "^0.8.0",
     "https-localhost": "^4.7.1",
-    "lint-staged": "^15.2.0",
+    "lint-staged": "^15.2.2",
     "markdown-it-link-attributes": "^4.0.1",
-    "markdown-it-shikiji": "^0.9.12",
-    "pnpm": "^8.13.1",
+    "markdown-it-shikiji": "^0.9.19",
+    "pnpm": "^8.15.6",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "rollup": "^4.9.1",
-    "shikiji": "^0.9.12",
-    "simple-git-hooks": "^2.9.0",
+    "rollup": "^4.14.0",
+    "shikiji": "^0.9.19",
+    "simple-git-hooks": "^2.11.1",
     "storybook": "^7.6.17",
-    "taze": "^0.13.1",
-    "typescript": "^5.3.3",
-    "unocss": "^0.58.5",
-    "unplugin-auto-import": "^0.17.3",
+    "taze": "^0.13.3",
+    "typescript": "^5.4.3",
+    "unocss": "^0.58.9",
+    "unplugin-auto-import": "^0.17.5",
     "unplugin-vue-components": "^0.26.0",
-    "unplugin-vue-macros": "^2.7.7",
+    "unplugin-vue-macros": "^2.9.0",
     "unplugin-vue-markdown": "^0.25.2",
     "unplugin-vue-router": "^0.7.0",
-    "vite": "^5.0.10",
-    "vite-bundle-visualizer": "^1.0.0",
-    "vite-plugin-inspect": "^0.8.1",
-    "vite-plugin-pwa": "^0.17.4",
-    "vite-plugin-vue-devtools": "^1.0.0-rc.8",
+    "vite": "^5.2.8",
+    "vite-bundle-visualizer": "^1.1.0",
+    "vite-plugin-inspect": "^0.8.3",
+    "vite-plugin-pwa": "^0.17.5",
+    "vite-plugin-vue-devtools": "1.0.0-rc.8",
     "vite-plugin-vue-layouts": "^0.11.0",
-    "vite-plugin-webfont-dl": "^3.9.1",
+    "vite-plugin-webfont-dl": "^3.9.2",
     "vite-ssg": "^0.23.6",
     "vite-ssg-sitemap": "^0.6.1",
-    "vitest": "^1.1.0",
+    "vitest": "^1.4.0",
     "vue-tsc": "^1.8.27"
   },
   "simple-git-hooks": {
